Main-Exam UI rewritten using Grids
Before it was implemented using horizontal and vertical flexboxes. But
I wanted to get familiar with grids and they just seemed to be a better
choice for the current use case.
No more unnecessary divs anymore, code is kinda cleaner, but don't know
about correctness.
As usual, I'm expecting so many bugs!
I should spend some more time on this.