Major work - Login / Signup
Initially, I thought I should have only one global state, and I should
do data-fetching in the redux-store itself (with redux-thunk). But it's
not efficient, because I am not using the data anywhere outside the
component.

That's why restructured the whole data-fetching, such that all the api
requests will be done in a component, and redux store will only contain,
well, the global state.

Took a lot of time, because didn't (don't?) know nitty-gritties of
`useCallback()` and `useEffect()` and stuff.